+++ src/assets/js/taskchecker.js
@@ -1,0 +1,18 @@
+import Tasks from './tasklist';
+
+const tasks = new Tasks();
+const taskchecker = () => {
+  const checkboxtf = document.querySelectorAll('input[type=checkbox]');
+  checkboxtf.forEach((box) => {
+    box.addEventListener('change', () => {
+      if (box.checked) {
+        tasks.list[box.id].completed = true;
+        localStorage.setItem('tasks', JSON.stringify(tasks.list));
+      } else {
+        tasks.list[box.id].completed = false;
+        localStorage.setItem('tasks', JSON.stringify(tasks.list));
+      }
+    });
+  });
+};
+export default taskchecker;
